refactored case view
