+++ src/views/auth/Login.vue
@@ -57,7 +57,7 @@
 
 <script>
 // @ is an alias to /src
-import LoginForm from "@/components/LoginForm.vue";
+import LoginForm from "@/components/auth/LoginForm.vue";
 export default {
   name: "Login",
   components: { LoginForm,  },
