+++ src/views/auth/Register.vue
@@ -60,7 +60,7 @@
 
 <script>
 // @ is an alias to /src
-import RegisterForm from "@/components/RegisterForm.vue";
+import RegisterForm from "@/components/auth/RegisterForm.vue";
 export default {
   name: "Register",
   components: { RegisterForm, },
