+++ src/views/protected/Settings.vue
@@ -13,7 +13,7 @@
 <script>
 // @ is an alias to /src
 
-import SettingsPage from '../../components/SettingsPage.vue';
+import SettingsPage from '../../components/auth/SettingsPage.vue';
 export default {
   name: "Settings",
   components: { SettingsPage },
